Working triggerable script
Tested and working. I merged the concept of effect in a unique script, this way connecting an effect to its triggers is easier as it only requires to setup a branch in scene tree.
It may also be used to trigger multiple effects in chain (to be tested).
I may inprove it by replacing getters and setters with methods so separating active variable from them I may also  make a reset method useful if I need to add a timer effect