Improved door movement
Tested hierarchical puzzle triggers. They work well but NOT needs to be fixed. It is not initialized well. After being turned on it works properly so it has to do with initialization, probably of the NOT trigger that should be set its own active state according to actual state on screen. If toch is off, it has to be active.